Finish Suppression process

+++ suppression.php
@@ -4,15 +4,19 @@ require_once __DIR__ . '/layout/header.php';
 
 require_once __DIR__ . '/classes/config.php';
 
-session_start();
+if (isset($_POST['user_id']) && isset($_POST['location'])) {
 
-$user_id = $_SESSION['user_id'];
+    echo $_POST['user_id'];
+    $sql = "DELETE FROM users WHERE Id_Users = :id";
+    $stmt = $pdo->prepare($sql);
+    $stmt->execute(['id' => $_POST['user_id']]);
 
-$sql = "DELETE FROM users WHERE Id_Users = :id";
-$stmt = $pdo->prepare($sql);
-$stmt->execute(['id' => $user_id]);
-$user = $stmt->fetch(PDO::FETCH_ASSOC);
-unset($_SESSION['user_id']);
-session_destroy();
-
-header("deconnexion.php");
+    if ($_POST['location'] == "settings") {
+        unset($_SESSION['user_id']);
+        header("Location: deconnexion.php");
+        exit;
+    } else {
+        header("Location: adminDashboard.php");
+        exit;
+    }
+}
